+++ math/src/test/java/proeqfa/math/estimation/ThreeLogicValuesTest.java
@@ -10,9 +10,9 @@ import static org.junit.Assert.*;
 public class ThreeLogicValuesTest {
 
     /*warn! order is important*/
-    float[] VIEW1 = {2, 1, 0},
-            VIEW2 = {1, 0, -1},
-            VIEW3 = {1, 0.5f, 0};
+    double[] VIEW1 = {2d, 1d, 0d},
+            VIEW2 = {1d, 0d, -1d},
+            VIEW3 = {1d, 0.5d, 0d};
 
     public ThreeLogicValuesTest() {
     }
@@ -25,8 +25,8 @@ public class ThreeLogicValuesTest {
         System.out.println("getView1");
         ThreeLogicValues result = ThreeLogicValues.getView1();
         assertNotNull(result);
-        assertArrayEquals(new float[]{result.getMore(), result.getSame(), result.getLess()},
-                VIEW1, 0);
+        assertArrayEquals(new double[]{result.getMore(), result.getSame(), result.getLess()},
+                VIEW1, 0d);
         assertEquals(result.getMore(), VIEW1[0], 0);
         assertEquals(result.getLess(), VIEW1[2], 0);
         assertEquals(result.getSame(), VIEW1[1], 0);
@@ -41,7 +41,7 @@ public class ThreeLogicValuesTest {
         System.out.println("getView2");
         ThreeLogicValues result = ThreeLogicValues.getView2();
         assertNotNull(result);
-        assertArrayEquals(new float[]{result.getMore(), result.getSame(), result.getLess()},
+        assertArrayEquals(new double[]{result.getMore(), result.getSame(), result.getLess()},
                 VIEW2, 0);
         assertEquals(result.getMore(), VIEW2[0], 0);
         assertEquals(result.getLess(), VIEW2[2], 0);
@@ -56,7 +56,7 @@ public class ThreeLogicValuesTest {
         System.out.println("getView3");
         ThreeLogicValues result = ThreeLogicValues.getView3();
         assertNotNull(result);
-        float[] resultValues = {result.getMore(), result.getSame(), result.getLess()};
+        double[] resultValues = {result.getMore(), result.getSame(), result.getLess()};
         assertEquals(result.getMore(), VIEW3[0], 0);
         assertEquals(result.getLess(), VIEW3[2], 0);
         assertEquals(result.getSame(), VIEW3[1], 0);
@@ -69,9 +69,9 @@ public class ThreeLogicValuesTest {
     @Test
     public void testGetMore() {
         System.out.println("getMore");
-        float expResult = 5.0F;
+        double expResult = 5.0d;
         ThreeLogicValues instance = new ThreeLogicValues(expResult, 2, 1);
-        float result = instance.getMore();
+        double result = instance.getMore();
         assertEquals(expResult, result, 0.0);
 
     }
@@ -82,9 +82,9 @@ public class ThreeLogicValuesTest {
     @Test
     public void testGetLess() {
         System.out.println("getLess");
-        float expResult = 1.0F;
+        double expResult = 1.0d;
         ThreeLogicValues instance = new ThreeLogicValues(3, 2, expResult);
-        float result = instance.getLess();
+        double result = instance.getLess();
         assertEquals(expResult, result, 0.0);
     }
 
@@ -94,9 +94,9 @@ public class ThreeLogicValuesTest {
     @Test
     public void testGetSame() {
         System.out.println("getSame");
-        float expResult = 2.0F;
+        double expResult = 2.0d;
         ThreeLogicValues instance = new ThreeLogicValues(3, expResult, 1);
-        float result = instance.getSame();
+        double result = instance.getSame();
         assertEquals(expResult, result, 0.0);
     }
